refactor(GraphView): move CytoscapeRenderer into its own component
